+++ Element/DataManagerElement.php
@@ -17,7 +17,6 @@ use Symfony\Component\Translation\TranslatorInterface;
 /**
  * @author  Andriy Oblivantsev <eslider@gmail.com>
  * @todo: add http method for form item reload (modified data may change "sql"-type select options, in any schema)
- * @todo: supply reasonable defaults for "tableTranslation" option
  * @todo: support lazy-loading of single schema, to speed up initialization
  * @todo: verify file upload field interactions
  */