+++ CHANGELOG.md
@@ -19,6 +19,10 @@
 - Drop support hacks for legacy vis-ui versions <0.1.80
 - Replace vis-ui.js-built.js dependency (provided by abandonware robloach/component-installer) with individual vendor-sourced file references
 
+# 1.0.12
+- Fix multi-select data extraction (convert to scalar by joining with separator)
+- Fix radio groups initializing with no radio button checked
+
 # 1.0.11
 - Fix backend form browser text searchability through off-screen portions of "schemes" area
 - Fix backend form sizing